Add option to save CA key
QuickTLS was designed to generate certificates and throw away the
CA key to guarantee the integrity of the CA and that no new
certificates can be generated beyond the original set. This change
adds a flag to make it easier to develop CA use cases which manage
the integrity of the CA key.